pyproject.toml
[project]
name = "algoart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "noise>=1.2.2",
    "numpy>=2.0.2",
    "pillow>=11.1.0",
]

[tool.uv]
dev-dependencies = [
    "pylint>=3.3.3",
    "ruff>=0.9.2",
]
